backend/requirements.txt
#intial dependency   
fastapi
uvicorn

#database endpoint
sqlalchemy
psycopg2

#env. var
python-dotenv
